Locally mapped bands.  Need to change the function wrapper api slightly.
