+++ datacube_ows/config_utils.py
@@ -3,6 +3,7 @@ import os
 from importlib import import_module
 from typing import Mapping, Sequence
 
+from datacube_ows.config_toolkit import deepinherit
 from datacube_ows.ogc_utils import ConfigException
 
 
@@ -119,26 +120,6 @@ class OWSIndexedConfigEntry(OWSConfigEntry):
     def lookup_impl(cls, cfg, keyvals):
         raise NotImplementedError()
 
-
-def deepinherit(parent, child):
-    for k in parent:
-        if k in child:
-            if isinstance(child[k], dict):
-                # recurse dictionary
-                deepinherit(parent[k], child[k])
-            elif isinstance(child[k], str):
-                # Keep child's version of str
-                pass
-            else:
-                try:
-                    iter(child[k])
-                    # non-str iterable - append child to parent
-                    child[k] = parent[k] + child[k]
-                except TypeError:
-                    # Non-iterable - keep child's version
-                    pass
-        else:
-            child[k] = parent[k]
 
 class OWSExtensibleConfigEntry(OWSIndexedConfigEntry):
     def __init__(self, cfg, keyvals, global_cfg, *args, **kwargs):
